Address comments - reduce comments, merge the two hooks and then also use map
