--- mlflow/server/js/src/experiment-tracking/pages/experiment-evaluation-datasets-v2/hooks/usePersistedTableColumnWidths.ts
@@ -1,89 +1,0 @@
-import { useCallback, useMemo } from 'react';
-import { useLocalStorage } from '@databricks/web-shared/hooks';
-import type { ColumnSizingState, OnChangeFn } from '@tanstack/react-table';
-
-const COLUMN_WIDTH_STORAGE_KEY_PREFIX = 'mlflow.eval-datasets.column-widths';
-
-// Bumping this constant invalidates every stored entry across all users —
-// `useLocalStorage` keys are namespaced by version, so old keys become
-// unreachable. Bump only when the stored shape is incompatible (e.g., if we
-// ever switch from absolute px to proportions). No migration path is offered;
-// users get the bucket defaults on first load after a bump.
-const COLUMN_WIDTH_STORAGE_VERSION = 1;
-
-interface UsePersistedTableColumnWidthsParams {
-  experimentId: string;
-  datasetId: string;
-}
-
-interface PersistedColumnWidths {
-  columnSizing: ColumnSizingState; // controlled state for useReactTable
-  setColumnSizing: OnChangeFn<ColumnSizingState>; // plugs straight into onColumnSizingChange
-}
-
-/**
- * Filters out non-positive-finite entries so corrupt/old localStorage values
- * (NaN, strings, null, negatives, 0) can't flow into `flex: 0 0 ${px}px` and
- * collapse a column. Always returns a fresh object — identity stability across
- * renders is the caller's job via `useMemo([stored])`, which gates this call
- * on the `stored` reference returned by `useLocalStorage`.
- */
-const sanitizeStoredWidths = (stored: unknown): ColumnSizingState => {
-  if (stored === null || typeof stored !== 'object') return {};
-  const entries = Object.entries(stored as Record<string, unknown>).filter(
-    ([, value]) => typeof value === 'number' && Number.isFinite(value) && value > 0,
-  );
-  return Object.fromEntries(entries) as ColumnSizingState;
-};
-
-/**
- * Persists column widths in localStorage, scoped per (experiment, dataset)
- * so each dataset can have its own remembered set. Starts empty (`{}`) and
- * only the columns the user actually drags get written — auto-defaults live
- * on the consumer (`COLUMN_WIDTHS_BY_BREAKPOINT` in `DatasetRecordsTable`).
- *
- * Why localStorage: synchronous read on mount avoids a column-flicker on
- * first paint; IndexedDB would require an async read. Sizes are tiny
- * (<200B per dataset).
- */
-export const usePersistedTableColumnWidths = ({
-  experimentId,
-  datasetId,
-}: UsePersistedTableColumnWidthsParams): PersistedColumnWidths => {
-  const key = `${COLUMN_WIDTH_STORAGE_KEY_PREFIX}.${experimentId}.${datasetId}`;
-
-  const [stored, setStored] = useLocalStorage<ColumnSizingState>({
-    key,
-    version: COLUMN_WIDTH_STORAGE_VERSION,
-    initialValue: {},
-  });
-
-  // Sanitise before exposing — protects every reader from corrupt entries
-  // (NaN, strings, negatives) that the schema-less storage layer might admit.
-  // Memo by `stored` so a clean run of widths keeps a stable reference.
-  const columnSizing = useMemo(() => sanitizeStoredWidths(stored), [stored]);
-
-  // Sanitise on the write path too. TanStack's `onColumnSizingChange` uses the
-  // functional-updater form (`prev => ({ ...prev, [id]: nextWidth })`) on every
-  // drag tick; without this wrapper, `prev` would be the *raw* stored value,
-  // so any corrupt entry already in localStorage would be spread into the new
-  // object and re-persisted alongside the user's drag — invisible to the read
-  // side, but lingering forever in storage. By feeding the updater a sanitized
-  // `prev` (and sanitising again before persist for the direct-value form), a
-  // corrupt entry dies on the next write instead of surviving indefinitely.
-  const setColumnSizing: OnChangeFn<ColumnSizingState> = useCallback(
-    (updater) => {
-      setStored((rawPrev) => {
-        const cleanPrev = sanitizeStoredWidths(rawPrev);
-        const next = typeof updater === 'function' ? updater(cleanPrev) : updater;
-        return sanitizeStoredWidths(next);
-      });
-    },
-    [setStored],
-  );
-
-  return {
-    columnSizing,
-    setColumnSizing,
-  };
-};
